# openapi.yaml
openapi: 3.1.0
info:
  title: HHC Web API
  version: 0.2.0
servers:
  - url: https://www.alive.org.tw/api
paths:
  /bulletins:
    get:
      operationId: listPublicBulletins
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Published bulletin page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicBulletinListEnvelope' }
  /bulletins/latest:
    get:
      operationId: getLatestPublicBulletin
      parameters: [{ $ref: '#/components/parameters/Locale' }]
      responses:
        '200': { $ref: '#/components/responses/PublicBulletin' }
        '404': { $ref: '#/components/responses/Error' }
  /bulletins/{issueDate}:
    get:
      operationId: getPublicBulletinByDate
      parameters:
        - { name: issueDate, in: path, required: true, schema: { type: string, format: date } }
        - $ref: '#/components/parameters/Locale'
      responses:
        '200': { $ref: '#/components/responses/PublicBulletin' }
        '404': { $ref: '#/components/responses/Error' }
...
  /admin/bulletins:
    get:
      operationId: listAdminBulletins
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/BulletinStatus' }
      responses:
        '200':
          description: CMS bulletin list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulletinListEnvelope' }
    post:
      operationId: createBulletinIssue
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBulletinInput' }
      responses:
        '201': { $ref: '#/components/responses/BulletinIssue' }
        '409': { $ref: '#/components/responses/Error' }
  /admin/bulletins/{issueId}:
    get:
      operationId: getAdminBulletin
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:read]
      parameters: [{ $ref: '#/components/parameters/IssueID' }]
      responses:
        '200': { $ref: '#/components/responses/BulletinIssue' }
        '404': { $ref: '#/components/responses/Error' }
  /admin/bulletins/{issueId}/upload-sessions:
    post:
      operationId: createBulletinUploadSession
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write, assets:write]
      parameters:
        - $ref: '#/components/parameters/IssueID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBulletinUploadInput' }
      responses:
        '201':
          description: Short-lived asset upload target created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedUploadEnvelope' }
  /admin/bulletins/{issueId}/assets/{assetId}/complete:
    post:
      operationId: completeBulletinUpload
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write, assets:write]
      parameters:
        - $ref: '#/components/parameters/IssueID'
        - { name: assetId, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteBulletinUploadInput' }
      responses:
        '200': { $ref: '#/components/responses/BulletinIssue' }
        '412': { $ref: '#/components/responses/Error' }
  /admin/bulletins/{issueId}/publish:
    post:
      operationId: publishBulletin
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:publish]
      parameters:
        - $ref: '#/components/parameters/IssueID'
        - $ref: '#/components/parameters/IfMatch'
      requestBody: { $ref: '#/components/requestBodies/PublicationInput' }
      responses:
        '202':
          description: Publication workflow accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkflowEnvelope' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/bulletins/{issueId}/unpublish:
    post:
      operationId: unpublishBulletin
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:publish]
      parameters:
        - $ref: '#/components/parameters/IssueID'
        - $ref: '#/components/parameters/IfMatch'
      requestBody: { $ref: '#/components/requestBodies/PublicationInput' }
      responses:
        '200': { $ref: '#/components/responses/BulletinIssue' }
        '412': { $ref: '#/components/responses/Error' }
...
  /admin/content/news/{contentId}/upload-sessions:
    post:
      operationId: createNewsCoverUpload
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write, assets:write]
      parameters:
        - $ref: '#/components/parameters/ContentID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateImageUploadInput' }
      responses:
        '201':
          description: News cover upload target
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedUploadEnvelope' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Locale: { name: locale, in: query, schema: { $ref: '#/components/schemas/Locale' } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    IssueID: { name: issueId, in: path, required: true, schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^"[1-9][0-9]*"$' } }
...
    ContentID: { name: contentId, in: path, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
    PublicationInput:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PublicationInput' }
...
  responses:
    BulletinIssue:
      description: Bulletin issue
      headers:
        ETag: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BulletinIssueEnvelope' }
    PublicBulletin:
      description: Published bulletin
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PublicBulletinEnvelope' }
    Error:
      description: Standard API error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Locale:
      type: string
      enum: [zh-Hant, zh-Hans, en]
      default: zh-Hant
    BulletinStatus:
      type: string
      enum: [draft, publishing, published, unpublishing, unpublish_failed, unpublished, archived]
    BulletinVersionStatus:
      type: string
      enum: [draft, publishing, published, unpublishing, unpublish_failed, unpublished]
...
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    BulletinVersion:
      type: object
      required: [id, issueId, locale, title, pdfAssetId, pdfFileName, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        issueId: { type: string, format: uuid }
        locale: { $ref: '#/components/schemas/Locale' }
        title: { type: string }
        pdfAssetId: { type: string }
        pdfFileName: { type: string }
        publicGrantId: { type: string }
        status: { $ref: '#/components/schemas/BulletinVersionStatus' }
        workflowStatus: { type: string }
        workflowError: { type: string }
        version: { type: integer, format: int64 }
        publishedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BulletinIssue:
      type: object
      required: [id, issueDate, status, version, createdBy, updatedBy, createdAt, updatedAt, versions]
      properties:
        id: { type: string, format: uuid }
        issueDate: { type: string, format: date }
        status: { $ref: '#/components/schemas/BulletinStatus' }
        version: { type: integer, format: int64 }
        createdBy: { type: string }
        updatedBy: { type: string }
        publishedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        versions:
          type: array
          items: { $ref: '#/components/schemas/BulletinVersion' }
    PublicBulletin:
      type: object
      required: [issueDate, locale, title, downloadUrl, publishedAt, version]
      properties:
        issueDate: { type: string, format: date }
        locale: { $ref: '#/components/schemas/Locale' }
        title: { type: string }
        downloadUrl: { type: string }
        publishedAt: { type: string, format: date-time }
        version: { type: integer, format: int64 }
    PublicBulletinIssue:
      type: object
      required: [issueDate, versions]
      properties:
        issueDate: { type: string, format: date }
        versions:
          type: array
          items: { $ref: '#/components/schemas/PublicBulletin' }
    Workflow:
      type: object
      required: [id, status, aggregateVersion, createdAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        aggregateVersion: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
    AssetSummary:
      type: object
      required: [id]
      properties:
        id: { type: string }
    UploadTarget:
      type: object
      required: [url, method, headers, expiresAt]
      properties:
        url: { type: string, format: uri }
        method: { type: string }
        headers: { type: object, additionalProperties: { type: string } }
        expiresAt: { type: string, format: date-time }
    CreatedUpload:
      type: object
      required: [asset, uploadTarget]
      properties:
        asset: { $ref: '#/components/schemas/AssetSummary' }
        uploadTarget: { $ref: '#/components/schemas/UploadTarget' }
    CreateBulletinInput:
      type: object
      additionalProperties: false
      required: [issueDate]
      properties:
        issueDate: { type: string, format: date }
    CreateBulletinUploadInput:
      type: object
      additionalProperties: false
      required: [locale, fileName, mimeType, sizeBytes]
      properties:
        locale: { $ref: '#/components/schemas/Locale' }
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, const: application/pdf }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 20971520 }
    CompleteBulletinUploadInput:
      type: object
      additionalProperties: false
      required: [locale, title, fileName, mimeType, sizeBytes, checksumSha256]
      properties:
        locale: { $ref: '#/components/schemas/Locale' }
        title: { type: string, minLength: 1 }
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, const: application/pdf }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 20971520 }
        checksumSha256: { type: string, pattern: '^[a-fA-F0-9]{64}$' }
    CreateImageUploadInput:
      type: object
      additionalProperties: false
      required: [fileName, mimeType, sizeBytes]
      properties:
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 10485760 }
...
    PublicationInput:
      type: object
      additionalProperties: false
      required: [locale]
      properties:
        locale: { $ref: '#/components/schemas/Locale' }
...
    BulletinIssueEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/BulletinIssue' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    BulletinListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/BulletinIssue' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
        error: { type: ['null'] }
    PublicBulletinEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/PublicBulletin' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    PublicBulletinListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/PublicBulletinIssue' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
        error: { type: ['null'] }
    CreatedUploadEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/CreatedUpload' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    WorkflowEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Workflow' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    ErrorEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data: { type: ['null'] }
        meta: { type: object, additionalProperties: true }
        error: { $ref: '#/components/schemas/ErrorBody' }
...